issue-bot-project/issue_bot/requirements.txt
asgiref==3.4.1
blis==0.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
ChatterBot==1.0.8
chatterbot-corpus==1.2.0
click==8.0.3
cymem==2.0.6
Deprecated==1.2.13
Django==2.2
djangorestframework==3.12.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz
future==0.18.2
idna==2.7
importlib-metadata==4.8.2
joblib==1.1.0
mathparse==0.1.2
murmurhash==1.0.6
nltk==3.6.5
numpy==1.21.4
oauthlib==3.1.1
packaging==21.3
Pint==0.18
plac==0.9.6
preshed==2.0.1
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
pymongo==3.12.1
PyNaCl==1.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-twitter==3.5
pytz==2021.3
PyYAML==3.13
regex==2021.11.10
requests==2.19.1
requests-oauthlib==1.3.0
six==1.16.0
spacy==2.1.9
SQLAlchemy==1.3.24
sqlparse==0.4.2
srsly==1.0.5
thinc==7.0.8
tqdm==4.62.3
typing-extensions==4.0.0
urllib3==1.23
wasabi==0.8.2
wrapt==1.13.3
zipp==3.6.0
stop_words==2018.7.23